feat: refactor exception handling classes to improve package structure
